style: templates RENAME
